+++ packages/hoppscotch-data/src/graphql/v/8.ts
@@ -9,10 +9,10 @@ import {
 } from "./2"
 import { HoppGQLAuthAPIKey } from "./4"
 import { HoppGQLAuthAWSSignature } from "./6"
-import { HoppRESTAuthOAuth2 } from "./../../rest/v/10"
+import { HoppRESTAuthOAuth2 } from "./../../rest/v/11"
 import { V7_SCHEMA } from "./7"
 
-export { HoppRESTAuthOAuth2 as HoppGQLAuthOAuth2 } from "../../rest/v/10"
+export { HoppRESTAuthOAuth2 as HoppGQLAuthOAuth2 } from "../../rest/v/11"
 
 export const HoppGQLAuth = z
   .discriminatedUnion("authType", [
@@ -43,28 +43,22 @@ export default defineVersion({
   up(old: z.infer<typeof V7_SCHEMA>) {
     const auth = old.auth
 
-    if (
-      auth.authType === "oauth-2" &&
-      auth.grantTypeInfo.grantType === "CLIENT_CREDENTIALS"
-    ) {
-      return {
-        ...old,
-        v: 8 as const,
-        auth: {
-          ...auth,
-          grantTypeInfo: {
-            ...auth.grantTypeInfo,
-            grantType: "CLIENT_CREDENTIALS" as const,
-            sendAs: "IN_BODY" as const,
-          },
-        },
-      }
-    }
-
     return {
       ...old,
-      auth: auth as HoppGQLAuth,
       v: 8 as const,
+      auth:
+        auth.authType === "oauth-2"
+          ? {
+              ...auth,
+              grantTypeInfo:
+                auth.grantTypeInfo.grantType === "CLIENT_CREDENTIALS"
+                  ? {
+                      ...auth.grantTypeInfo,
+                      sendAs: "IN_BODY" as const,
+                    }
+                  : auth.grantTypeInfo,
+            }
+          : auth,
     }
   },
 })
