+++ packages/hoppscotch-data/src/rest/index.ts
@@ -16,7 +16,8 @@ import V6_VERSION from "./v/6"
 import V7_VERSION, { HoppRESTHeaders, HoppRESTParams } from "./v/7"
 import V8_VERSION from "./v/8"
 import V9_VERSION, { HoppRESTRequestResponses } from "./v/9"
-import V10_VERSION, { HoppRESTReqBody, HoppRESTAuth } from "./v/10"
+import V10_VERSION, { HoppRESTReqBody } from "./v/10"
+import V11_VERSION, { HoppRESTAuth } from "./v/11"
 
 export * from "./content-types"
 
@@ -56,7 +57,7 @@ export {
   HoppRESTAuthOAuth2,
   HoppRESTAuth,
   ClientCredentialsGrantTypeParams,
-} from "./v/10"
+} from "./v/11"
 
 const versionedObject = z.object({
   // v is a stringified number
@@ -64,7 +65,7 @@ const versionedObject = z.object({
 })
 
 export const HoppRESTRequest = createVersionedEntity({
-  latestVersion: 10,
+  latestVersion: 11,
   versionMap: {
     0: V0_VERSION,
     1: V1_VERSION,
@@ -77,6 +78,7 @@ export const HoppRESTRequest = createVersionedEntity({
     8: V8_VERSION,
     9: V9_VERSION,
     10: V10_VERSION,
+    11: V11_VERSION,
   },
   getVersion(data) {
     // For V1 onwards we have the v string storing the number
@@ -119,7 +121,7 @@ const HoppRESTRequestEq = Eq.struct<HoppRESTRequest>({
   responses: lodashIsEqualEq,
 })
 
-export const RESTReqSchemaVersion = "10"
+export const RESTReqSchemaVersion = "11"
 
 export type HoppRESTParam = HoppRESTRequest["params"][number]
 export type HoppRESTHeader = HoppRESTRequest["headers"][number]
